Moved startup out
